backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1          # passlib 1.7.4 tidak kompatibel dengan bcrypt>=4.1
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
pillow==10.3.0
python-dotenv==1.0.1
slowapi==0.1.9
email-validator==2.1.1
geopy==2.4.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
coverage==7.5.1

# Database drivers (choose one)
pymysql==1.1.1          # MySQL/MariaDB
# psycopg2-binary==2.9.9  # PostgreSQL (uncomment if using PG)

mangum==0.17.0
supabase==2.4.6
